[#69] Move TabItemTitle into file
