+++ NimbleMedium.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -24,7 +24,7 @@
         "repositoryURL": "https://github.com/markgravity/PagerTabStripView.git",
         "state": {
           "branch": "master",
-          "revision": "2bbf6d99a6bbea696abaaf1c9f9e6172863f7245",
+          "revision": "dd08c6cf0d18427c46a134f731b3f50410a8fdd7",
           "version": null
         }
       },
