+++ NimbleMedium/Sources/Constants/Constants+API.swift
@@ -9,5 +9,5 @@
 extension Constants.API {
 
     // For fixing performance issues and the presence of NSFW content on the current API, we encourage you to use the new temporary API: https://realworld-temp-api.herokuapp.com/api Please keep in mind we'll soon move back to the current API: https://conduit.productionready.io/api
-    static let baseURL: String = "https://realworld-temp-api.herokuapp.com/api"
+    static let baseURL: String = "https://conduit.productionready.io/api"
 }
