+++ NimbleMedium/Sources/Presentation/Modules/UserProfile/UserProfileView.swift
@@ -40,10 +40,8 @@ struct UserProfileView: View {
                     }
             }
             .pagerTabStripViewStyle(
-                .barButton(
-                    indicatorBarHeight: 2.0,
+                .normal(
                     indicatorBarColor: .green,
-                    tabItemSpacing: 0.0,
                     tabItemHeight: 50.0,
                     placedInToolbar: false
                 )
@@ -75,47 +73,6 @@ struct UserProfileView: View {
                 .lineLimit(2)
                 .multilineTextAlignment(.center)
                 .padding()
-        }
-    }
-}
-
-private extension UserProfileView {
-
-    class NavTabViewTheme: ObservableObject {
-        @Published var textColor = Color.gray
-    }
-
-    struct TabItemTitle: View, PagerTabViewDelegate {
-
-        let title: String
-
-        @ObservedObject fileprivate var theme = NavTabViewTheme()
-
-        var body: some View {
-            VStack {
-                Spacer()
-                Text(title)
-                    .foregroundColor(theme.textColor)
-                    .font(.subheadline)
-                Spacer()
-                Divider()
-                    .frame(height: 1)
-                    .background(Color(R.color.lightGray.name))
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-        }
-
-        init(_ title: String) {
-            self.title = title
-        }
-
-        func setState(state: PagerTabViewState) {
-            switch state {
-            case .selected:
-                theme.textColor = .green
-            default:
-                theme.textColor = .gray
-            }
         }
     }
